fix: move analysis module into analysis folder
