+++ docs/api/analysis/cross_shop.md
@@ -1,0 +1,3 @@
+# Cross Shop Analysis
+
+::: pyretailscience.analysis.cross_shop
