+++ docs/api/analysis/customer.md
@@ -1,0 +1,3 @@
+# Customer Analysis
+
+::: pyretailscience.analysis.customer
