+++ docs/api/analysis/gain_loss.md
@@ -1,0 +1,3 @@
+# Gain Loss Analysis
+
+::: pyretailscience.analysis.gain_loss
