+++ docs/api/analysis/product_association.md
@@ -1,0 +1,3 @@
+# Product Associations
+
+::: pyretailscience.analysis.product_association
