+++ docs/api/analysis/revenue_tree.md
@@ -1,0 +1,3 @@
+# Customer Analysis
+
+::: pyretailscience.analysis.revenue_tree
