+++ docs/api/analysis/segmentation.md
@@ -1,0 +1,3 @@
+# Segmentation
+
+::: pyretailscience.analysis.segmentation
