--- docs/api/cross_shop.md
@@ -1,3 +1,0 @@
-# Cross Shop Analysis
-
-::: pyretailscience.cross_shop
