--- docs/api/customer.md
@@ -1,3 +1,0 @@
-# Customer Analysis
-
-::: pyretailscience.customer
