--- docs/api/gain_loss.md
@@ -1,3 +1,0 @@
-# Gain Loss Analysis
-
-::: pyretailscience.gain_loss
