--- docs/api/product_association.md
@@ -1,3 +1,0 @@
-# Product Associations
-
-::: pyretailscience.product_association
