--- docs/api/revenue_tree.md
@@ -1,3 +1,0 @@
-# Customer Analysis
-
-::: pyretailscience.revenue_tree
