--- docs/api/segmentation.md
@@ -1,3 +1,0 @@
-# Segmentation
-
-::: pyretailscience.segmentation
