+++ pyretailscience/analysis/gain_loss.py
